Revert "Fix random maintenance (#192)"
This reverts commit d84d7eb3e589bee114ee78c124c34dd782120a1c.

+++ code/game/objects/effects/spawners/roomspawner.dm
@@ -9,6 +9,10 @@
 	var/room_width = 0
 	var/room_height = 0
 
+/obj/effect/spawner/room/proc/LateSpawn()
+	template.load(get_turf(src), centered = template.centerspawner)
+	qdel(src)
+
 /obj/effect/spawner/room/Initialize()
 	..()
 	return INITIALIZE_HINT_LATELOAD
@@ -29,11 +33,11 @@
 		template.weight = (template.weight / 2)
 		if(template.stock <= 0)
 			template.spawned = TRUE
-		template.load(get_turf(src), centered = template.centerspawner)
+		addtimer(CALLBACK(src, /obj/effect/spawner/room.proc/LateSpawn), 600)
+	else 
+		template = null
+	if(!template)
 		qdel(src)
-	else
-		qdel(src)
-		CRASH("couldn't find template for a room spawner")
 
 /obj/effect/spawner/room/fivexfour
 	name = "5x4 room spawner"
